chore: resolve threads in pr part 1
