+++ src/app/[locale]/(default)/rules/(main)/loading.tsx
@@ -1,5 +1,9 @@
 import { RuleCardListSkeleton } from '@/components/rules/RuleCardListSkeleton';
 
 export default function RulesSkeleton() {
-  return <RuleCardListSkeleton />;
+  return (
+    <div className='mt-1 mb-1'>
+      <RuleCardListSkeleton />
+    </div>
+  );
 }
